moved emailing to a new component
