+++ app/Components/SendEmailContainer.js
@@ -1,0 +1,8 @@
+import { connect } from 'react-redux'
+import SendEmail from './SendEmail'
+
+const mapStateToProps = (state) => {
+  return state
+}
+
+export default connect(mapStateToProps, null)(SendEmail)
